Implemented gnu extern inline function definitions such that you can get both the optimized behavior and unoptimized behavior.
Optimized is to treat them as weak static inline functions.
Unoptimized is to treat them as extern prototypes; the choice between
these two is selected by the tracing flag
handleExternInline-asPrototype.

All of this special treatment for extern inline depends on the CCLang
flag handleExternInlineSpecially being set; it is changed from
treatExternInlineAsPrototype which only caused the as-prototype
behavior to be turned on.

Note that the consequence of tracing flags being off by default is
that the optimized behavior is now the default.