Various changes so we can combine different stages in Oink/Qual and re-use the traversal done by the XmlWriter.  This class should be renamed later.
XmlWriter and subclasses now take 1) a IdentityManager and 2) a
SerializeOracle.

If the output stream, which is now a pointer, is NULL, then no writing is done.

IdentityManager encapsulates the uniqueId and 'printed' functions and data, to
make XmlWriter re-entrant.

SerializeOracle has a bunch of shouldSerialize() virtual functions to query
whether to serialize this item.  These need to be subclassable so that when
doing the 'mark externs' pass in Qual we can traverse every node (we're the
ones SETTING whether to serialize later).

In XmlWriter, new class XmlTagPrinter replaces XmlCloseTagPrinter and
encapsulates IncDec also.  It has helper methods which replace the "openTag*"
macros.  XmlWriter.writingP should be called to check whether we're actually
writing or just doing traversal.

More efficient indentation printing.