Removed the TranslationUnit* argument to the Env and Variable ctors as Oink no longer needs it.
Generalized class MarkRealVars and markRealVariables() to visit
instead of mark so they are reusable.

The Env member madeUpVariables is now a reference; the client must
make one and pass it in.  This was done because Oink (and probably
other analyses) needs to visit these variables after typechecking is
done (which was the whole point of maintaining the list of them); as a
member of Env they would just go away at the end of typechecking.

No longer consider variables to have linkerVisibleName() if they
hasFlag(DF_NAMESPACE) or their names start with "__builtin_va".