Added a traversal of Scopes to the RealVarAndTypeASTVisitor class and its helper classes.  This was needed as not all real (non-template) Variables were being visited by the "real" variable visitation.
Annotated TranslationUnit AST nodes with a pointer to their global
scope.  I am using this for the scope traversal.  Added a
TranslationUnit argument to the ctor of the typechecking environment
class Env; it seems rather natural to me that the typechecking
environment should know what translation unit it is typechecking.