New Variable accessor functions: inGlobalOrNamespaceScope(), isStaticLinkage(), isStaticLocal(), isStaticMember(), isNonStaticMember().
These check various flags to compute do what they say they compute.

isStatic() is obsoleted; use the other isStatic*() functions.

In C++ mode, isStaticLinkage() now returns true for functions declared
"inline" without either "extern" or "static".  Note that this is *not* true in
C.  This includes static members, which now also have static linkage.  Thus
Variable::linkerVisibleName() now returns false for these.

New Env::inlineImpliesStaticLinkage which controls this behavior and is on
only in C++.

Static is now allowed to propagate to prototypes if it was implied by an "inline".