Checking in a patch from Scott who can't get svn commit to work right now due to password problems. ___ Subject: Re: the one elsa error that is left holding back Mozilla and libstdc++    Date: Sun, 17 Sep 2006 14:01:52 -0700 (PDT)    From: Scott McPeak <smcpeak@EECS.Berkeley.EDU>      To: "Daniel S. Wilkerson" <dsw@EECS.Berkeley.EDU>
Attached is a diff that "fixes" the testcase by disabling 
syncDefaultArgsWithDefinition() altogether.  See the comments for more 
info.

-Scott
___
Subject: Re: the one elsa error that is left holding back Mozilla and libstdc++
   Date: Sun, 17 Sep 2006 14:11:55 -0700 (PDT)
   From: Scott McPeak <smcpeak@EECS.Berkeley.EDU>
     To: "Daniel S. Wilkerson" <dsw@EECS.Berkeley.EDU>

No, the default arguments are used, they are attached to the declaration. 
This function was for some reason copying them to the definition too. 
I can't think of any reason for this, nor do any of my tests reveal the 
need for it.

It all seems to work, again w/o any copying, if the declaration and 
definition are the same.  Again, see the comments.

-Scott

On Sun, 17 Sep 2006, Daniel S. Wilkerson wrote:

> Does that mean that the default arugments are not used and so it is not
> analyzing C++ correctly?
>
> Daniel