New test case in/d0128.cc from from Umesh Shankar at Google <ushankar@google.com>.
This was a problem with initializers to pointers to members of
template type parameters within a template primary.  In general
whenever we call elaborateImplicitConversionArgToParam() we should
wrap it with a check that the first argument type does not
containsGeneralizedDependent().

I fixed this by adding the wrapper test as above.  I also changed the
other place where elaborateImplicitConversionArgToParam() was called
to check containsGeneralizedDependent() instead of just
isGeneralizedDependent(); I think this is a stronger test now and is
the right one.