Qualifiers (such as 'const' and 'volitile') are not allowed on references, and so when the list of qualifiers was extended for C99 to include 'restrict', that list did not automatically apply to references.  However the gnu people have taken it upon themselves to allow the 'restrict' qualifier on references as well, even though my guess is that it is not in the standard because Scott refers to 'restrict' as a C99 language-aspect throughout.
Therefore I had to extend the grammar slightly to allow for this,
though any qualifiers on references are currently dropped on the
floor.  For symmetry in the grammar I allow any qualifiers on a
reference, even though gcc does not accept 'const' there (though Scott
notes in one of his error tests that the Intel compiler does accept
it; I had to turn that error test off).

The test in/k0104.cc now passes.