Function implconv.h:getImplicitConversion() takes arguments which include a 'SpecialExpr special' which is the special properties of the source expression in the conversion attempt.
However cc_tcheck.cc:attemptCondConversion() did not take an argument
for the SpecialExpr of the expression corresponding to Type *t1 so it
could not pass the correct special to getImplicitConversion(); it was
instead passing SE_NONE, which is incorrect in the special cases where
it matters.

Now attemptCondConversion() takes a SpecialExpr special1 argument
which is computed to be the special of the same expression that
provides the type for t1.  See the new API below.

// given two types, find an implicit conversion between them, or
// return IC_NONE if none exists (do *not* insert error messages
// into the environment, either way)
ImplicitConversion getImplicitConversion(
  Env &env,            // type checking environment
  SpecialExpr special, // properties of the source expression
  Type *src,           // source type
  Type *dest,          // destination type
  bool destIsReceiver = false    // true if destination type is to be receiver object for method call
);

// try to convert t1 to t2 as per the rules in 5.16 para 3; return
// NULL if conversion is not possible, otherwise return type to which
// 't1' is converted (it is not always exactly 't2') and set 'ic'
// accordingly
Type *attemptCondConversion(Env &env, ImplicitConversion &ic /*OUT*/,
                            Type *t1, Type *t2,
                            SpecialExpr special1 // dsw: the special for t1
                            ) {